requirements.txt
nba_api>=1.4
duckdb>=0.9.0
pandas>=2.0
pyarrow>=14.0
numpy>=1.24
torch>=2.0
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43
captum>=0.6
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65
joblib>=1.3
pytest>=9.0